Move Log into hlt directory.
