+++ starter_kits/Swift/MyBot.swift
@@ -32,34 +32,6 @@ struct Constant: Codable {
     let game_seed: Int
 }
 
-class Log {
-    let filename: String
-    private let logFile : FileHandle
-    
-    init(filename: String) {
-        self.filename = filename
-        logFile = FileHandle(forWritingAtPath: filename)!
-    }
-    deinit {
-        logFile.closeFile()
-    }
-    
-    func debug(_ string: String) {
-        append(prefix: "DEBUG", message: string)
-    }
-    func info(_ string: String) {
-        append(prefix: "INFO", message: string)
-    }
-    func error(_ string: String) {
-        append(prefix: "ERROR", message: string)
-    }
-    
-    private func append(prefix: String, message: String) {
-        let string = "[\(prefix)] \(message)"
-        logFile.write(string.data(using: .utf8)!)
-    }
-}
-
 class Game {
     var turnNumber = 0
     let log = Log(filename: "output.log")
