+++ starter_kits/Swift/run_game.sh
@@ -2,5 +2,5 @@
 
 set -e
 
-swiftc MyBot.swift
+swiftc MyBot.swift hlt/*.swift
 ./halite --replay-directory replays/ -vvv --width 32 --height 32 "./MyBot" "./MyBot"
